requirements/base.txt
django==1.8.9
six==1.10.0
djangorestframework==3.3.2
markdown==2.6.2
django-filter==0.11.0
django-url-filter==0.2.0
django-dirtyfields==1.0
pymongo==2.8.1
mongoengine==0.13.0
blinker==1.4
django-rest-framework-mongoengine==3.3.1
beautifulsoup4==4.6.0
